Add an Octane version blacklist.

+++ manage.py
@@ -9,6 +9,8 @@ if __name__ == "__main__":
 
     from django.core.management import execute_from_command_line
 
+    version_blacklist = ['0.8.0']
+
     # Ensure Octane is kept up-to-date.
     if 'runserver' in sys.argv[1]:
         octane_release = requests.get('https://api.github.com/repos/tfausak/octane/releases/latest', headers={
@@ -31,20 +33,23 @@ if __name__ == "__main__":
         print('GH: {}. RLR: {}'.format(octane_release['name'], current_version))
 
         if tuple(octane_release['name'].split('.')) > tuple(current_version.split('.')):
-            # Download the latest version.
-            for file in current_binaries:
-                os.remove('octane-binaries/{}'.format(file))
+            if octane_release['name'] in version_blacklist:
+                print('Skipping this version.')
+            else:
+                # Download the latest version.
+                for file in current_binaries:
+                    os.remove('octane-binaries/{}'.format(file))
 
-            for asset in octane_release['assets']:
-                if 'windows' in asset['browser_download_url']:
-                    continue
+                for asset in octane_release['assets']:
+                    if 'windows' in asset['browser_download_url']:
+                        continue
 
-                print('Downloading {}'.format(asset['name']))
+                    print('Downloading {}'.format(asset['name']))
 
-                os.system('wget -qP octane-binaries/ {} && gunzip -f octane-binaries/{} && chmod +x octane-binaries/{}'.format(
-                    asset['browser_download_url'],
-                    asset['name'],
-                    asset['name'].replace('.gz', ''),
-                ))
+                    os.system('wget -qP octane-binaries/ {} && gunzip -f octane-binaries/{} && chmod +x octane-binaries/{}'.format(
+                        asset['browser_download_url'],
+                        asset['name'],
+                        asset['name'].replace('.gz', ''),
+                    ))
 
     execute_from_command_line(sys.argv)
